src/requirements.txt
pillow>=10.0.0
numpy>=1.24.0
pyvips>=2.2.0
pydicom>=2.4.0
